BUG: Update launcher settings removing broken method of gdb usage.
Users should attach to a running Slicer using the method described here:
https://www.slicer.org/wiki/Documentation/Nightly/Developers/Tutorials/Debug_Instructions#GDB_debug_by_attaching_to_running_process_.5BRECOMMENDED.5D

I assume that ddd usage is affected in the same was as gdb usage, as ddd
is a graphical wrapper around debuggers including gdb. To that end, this
form of ddd usage is removed as well.

From: James Hoctor <james.hoctor@kitware.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28772 3bd1e089-480b-0410-8dfb-8563597acbee